+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- madsim: Replace `SmallRng` with `Xoshiro256PlusPlus` for reproducibility across platforms.
+
 ## [0.2.17] - 2023-02-17
 
 ### Fixed